refactor: Switch to using SEA for creating standalong binaries
