--- src/img/do-not-delete
@@ -1,1 +1,0 @@
-...
