+++ src/lib/browser/__tests__/browser_complex_scenarios.test.js
