+++ src/lib/browser/__tests__/browser_edge_cases.test.js
