+++ src/lib/browser/__tests__/browser_install_uninstall.test.js
