+++ src/lib/browser/__tests__/browser_list_installed.test.js
