+++ src/lib/cloud/cloud-collections.js
@@ -2,7 +2,7 @@
 /* eslint-disable no-await-in-loop */
 
 const { table } = require('table');
-const { logger, setLoggingLevel, bsiExecutablePath, isPkg } = require('../../globals.js');
+const { logger, setLoggingLevel, bsiExecutablePath, isSea } = require('../../globals.js');
 const QlikSaas = require('./cloud-repo');
 const { qscloudTestConnection } = require('./cloud-test-connection');
 
@@ -30,7 +30,7 @@ const qscloudListCollections = async (options) => {
         setLoggingLevel(options.loglevel);
 
         logger.verbose('Starting listing of available collections');
-        logger.verbose(`Running as standalone app: ${isPkg}`);
+        logger.verbose(`Running as standalone app: ${isSea}`);
         logger.debug(`BSI executable path: ${bsiExecutablePath}`);
         logger.debug(`Options: ${JSON.stringify(options, null, 2)}`);
