+++ src/lib/qseow/qseow-enigma.js
@@ -1,10 +1,10 @@
-/* eslint-disable global-require */
-/* eslint-disable import/no-dynamic-require */
-const SenseUtilities = require('enigma.js/sense-utilities');
 const WebSocket = require('ws');
 const fs = require('fs-extra');
-const upath = require('upath');
+const SenseUtilities = require('enigma.js/sense-utilities');
+
 const { logger, bsiExecutablePath } = require('../../globals');
+const { getEnigmaSchema } = require('../util/enigma-util');
+const { getCertFilePaths } = require('../util/cert');
 
 /**
  *
@@ -35,16 +35,28 @@ const readCert = (filename) => fs.readFileSync(filename);
 const setupEnigmaConnection = (appId, options, command) => {
     logger.debug(`Prepping for QSEoW Enigma connection for app ${appId}`);
 
-    const certFile = upath.isAbsolute(options.certfile)
-        ? options.certfile
-        : upath.join(bsiExecutablePath, options.certfile);
-    const keyFile = upath.isAbsolute(options.certkeyfile)
-        ? options.certkeyfile
-        : upath.join(bsiExecutablePath, options.certkeyfile);
+    // Set up enigma.js configuration
+    logger.debug(`Enigma.js schema version: ${options.schemaversion}`);
 
-    // try {
-    const qixSchema = require(`enigma.js/schemas/${options.schemaversion}`);
-    logger.debug(`Successfully required Enigma`);
+    const qixSchema = getEnigmaSchema(options);
+    logger.debug(`Successfully loaded Enigma schema.`);
+
+    logger.verbose(`Using certificates for authentication with Enigma`);
+    logger.debug(`BSI executable path: ${bsiExecutablePath}`);
+
+    // Get certificate paths
+    const { fileCert, fileCertKey } = getCertFilePaths(options);
+
+    // const certFile = upath.isAbsolute(options.certfile)
+    //     ? options.certfile
+    //     : upath.join(bsiExecutablePath, options.certfile);
+    // const keyFile = upath.isAbsolute(options.certkeyfile)
+    //     ? options.certkeyfile
+    //     : upath.join(bsiExecutablePath, options.certkeyfile);
+
+    logger.debug(`Cert file: ${fileCert}`);
+    logger.debug(`Key file: ${fileCertKey}`);
+    // logger.debug(`CA file: ${fileCertCA}`);
 
     return {
         schema: qixSchema,
@@ -57,8 +69,9 @@ const setupEnigmaConnection = (appId, options, command) => {
         }),
         createSocket: (url) =>
             new WebSocket(url, {
-                key: readCert(keyFile),
-                cert: readCert(certFile),
+                key: readCert(fileCertKey),
+                cert: readCert(fileCert),
+                // ca: [readCert(fileCertCA)],
                 headers: {
                     'X-Qlik-User': `UserDirectory=${options.apiuserdir};UserId=${options.apiuserid}`,
                 },
@@ -66,9 +79,6 @@ const setupEnigmaConnection = (appId, options, command) => {
                     options.rejectUnauthorized === 'true' || options.rejectUnauthorized === true,
             }),
     };
-    // } catch (err) {
-    //     logger.error(`ENIGMA: ${err}`);
-    // }
 };
 
 module.exports = {
