+++ src/lib/qseow/qseow-remove-sheet-icons.js
@@ -1,10 +1,8 @@
-/* eslint-disable no-await-in-loop */
-/* eslint-disable import/extensions */
 const enigma = require('enigma.js');
 const qrsInteract = require('qrs-interact');
 
 const { setupEnigmaConnection } = require('./qseow-enigma.js');
-const { logger, setLoggingLevel, bsiExecutablePath, isPkg } = require('../../globals.js');
+const { logger, setLoggingLevel, bsiExecutablePath, isSea } = require('../../globals.js');
 const { qseowVerifyCertificatesExist } = require('./qseow-certificates.js');
 const { setupQseowQrsConnection } = require('./qseow-qrs.js');
 
@@ -147,7 +145,7 @@ const qseowRemoveSheetIcons = async (options) => {
         setLoggingLevel(options.loglevel);
 
         logger.info('Starting creation of thumbnails for Qlik Sense Enterprise on Windows (QSEoW)');
-        logger.verbose(`Running as standalone app: ${isPkg}`);
+        logger.verbose(`Running as standalone app: ${isSea}`);
         logger.debug(`BSI executable path: ${bsiExecutablePath}`);
         logger.debug(`Options: ${JSON.stringify(options, null, 2)}`);
